Refactor plugin files
