--- lib/bts_bot/echo.rb
@@ -1,20 +1,0 @@
-module BtsBot
-  class Echo
-    attr_reader :command, :description
-
-    def initialize
-      @command = :echo
-      @description = '入力した発言をそのままオウム返しします'
-    end
-
-    def set_callback(bot)
-      bot.command(@command, description: @description) do |e, str|
-        invoke(str)
-      end
-    end
-
-    def invoke(str)
-      str
-    end
-  end
-end
