+++ lib/bts_bot/plugin/echo.rb
@@ -1,0 +1,22 @@
+module BtsBot
+  module Plugin
+    class Echo
+      attr_reader :command, :description
+
+      def initialize
+        @command = :echo
+        @description = '入力した発言をそのままオウム返しします'
+      end
+
+      def set_callback(bot)
+        bot.command(@command, description: @description) do |e, str|
+          invoke(str)
+        end
+      end
+
+      def invoke(str)
+        str
+      end
+    end
+  end
+end
