--- lib/bts_bot/usagi.rb
@@ -1,20 +1,0 @@
-module BtsBot
-  class Usagi
-    attr_reader :command, :description
-
-    def initialize
-      @command = :usagi
-      @description = 'かわいい と返します (動作確認用)'
-    end
-
-    def set_callback(bot)
-      bot.command(@command, description: @description) do |e, str|
-        invoke
-      end
-    end
-
-    def invoke
-      'かわいい'
-    end
-  end
-end
